requirements.txt
llama-cpp-python>=0.2.44
huggingface_hub>=0.21.3
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.119
beautifulsoup4>=4.12.2
torch>=2.2.0
schedule>=1.2.0
